Support bzr's different merge algorithms
One can now run the tests for bzr with a merge algorithm other than
merge3.  For example, to run the tests with weave merge:

    $ env MERGE=weave make bzr

I hoped that one of the merge algorithms would be able to pass the
adjacent-changes test.  Unfortunately, they all produce the same overall
test results as using merge3.

Not all algorithms support '--show-base' so we use '--reprocess' to
minimize spurious conflicts in those cases.